En el modal de cantidad, agrega el mensaje de no disponibilidad de stock. Ya no muestra el mensaje en un modal aparte. En el componente de factura venta, quita los parentesis del header de Cerrar Venta. En el componente de factura venta, oculta la tabla de pagos cuando no tiene ninguno. En el componente de factura venta, ya no muestra el error al intentar agregar un primer pago. En el componente ver factura venta, todos los datos numéricos ya tienen formato. Remueve la key facturaDePedido cuando se hace click para facturar un pedido.
